pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xaitoolkit"
version = "0.1.0"
description = "Explain PyTorch image models with lightweight XAI methods (saliency, SmoothGrad, IG, RISE, LIME-stratified, axis-aligned SHAP-like)."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{name="Muhammad Rashid"}]
dependencies = [
  "torch",
  "torchvision",
  "timm",
  "numpy",
  "pillow",
  "matplotlib",
  "scikit-learn",
  "scikit-image",
]

[project.scripts]
xai-explain = "xaitoolkit.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
